+++ slashing-protection/src/test/java/tech/pegasys/web3signer/slashingprotection/validator/GenesisValidatorRootValidatorTest.java
@@ -13,6 +13,13 @@
 package tech.pegasys.web3signer.slashingprotection.validator;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 import tech.pegasys.web3signer.slashingprotection.dao.MetadataDao;
 
@@ -23,22 +30,93 @@ import org.jdbi.v3.core.Handle;
 import org.jdbi.v3.core.Jdbi;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Spy;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(DatabaseSetupExtension.class)
+@ExtendWith(MockitoExtension.class)
 class GenesisValidatorRootValidatorTest {
+  @Spy private MetadataDao metadataDao = new MetadataDao();
+
   @Test
   void checkGenesisValidatorReturnTrueForNewGVR(final Jdbi jdbi) {
     final GenesisValidatorRootValidator gvrValidator =
-        new GenesisValidatorRootValidator(jdbi, new MetadataDao());
-    assertThat(
-            gvrValidator.checkGenesisValidatorsRootAndInsertIfEmpty(Bytes32.leftPad(Bytes.of(3))))
+        new GenesisValidatorRootValidator(jdbi, metadataDao);
+    final Bytes32 genesisValidatorsRoot = Bytes32.leftPad(Bytes.of(3));
+    assertThat(gvrValidator.checkGenesisValidatorsRootAndInsertIfEmpty(genesisValidatorsRoot))
         .isTrue();
+  }
+
+  @Test
+  void verifyCachedGVRIsUsedForNewGVR(final Jdbi jdbi) {
+    final GenesisValidatorRootValidator gvrValidator =
+        new GenesisValidatorRootValidator(jdbi, metadataDao);
+    final Bytes32 gvr = Bytes32.leftPad(Bytes.of(3));
+    final Bytes32 otherGvr = Bytes32.leftPad(Bytes.of(4));
+
+    // call checkGVR multiple times, only first call should interact with database.
+    assertThat(gvrValidator.checkGenesisValidatorsRootAndInsertIfEmpty(gvr)).isTrue();
+    assertThat(gvrValidator.checkGenesisValidatorsRootAndInsertIfEmpty(gvr)).isTrue();
+    assertThat(gvrValidator.checkGenesisValidatorsRootAndInsertIfEmpty(otherGvr)).isFalse();
+
+    // verify database interactions
+    verify(metadataDao, times(1)).findGenesisValidatorsRoot(any());
+    verify(metadataDao, times(1)).insertGenesisValidatorsRoot(any(), eq(gvr));
+  }
+
+  @Test
+  void verifyCachedGVRIsUsedForExistingGVR(final Jdbi jdbi, final Handle handle) {
+    final GenesisValidatorRootValidator gvrValidator =
+        new GenesisValidatorRootValidator(jdbi, metadataDao);
+    final Bytes32 gvr = Bytes32.leftPad(Bytes.of(3));
+    final Bytes32 otherGvr = Bytes32.leftPad(Bytes.of(4));
+
+    insertGvr(handle, gvr);
+
+    // call checkGVR multiple times, only first call should interact with database.
+    assertThat(gvrValidator.checkGenesisValidatorsRootAndInsertIfEmpty(gvr)).isTrue();
+    assertThat(gvrValidator.checkGenesisValidatorsRootAndInsertIfEmpty(gvr)).isTrue();
+    assertThat(gvrValidator.checkGenesisValidatorsRootAndInsertIfEmpty(otherGvr)).isFalse();
+
+    // verify database interactions
+    verify(metadataDao, times(1)).findGenesisValidatorsRoot(any());
+    verify(metadataDao, never()).insertGenesisValidatorsRoot(any(), any());
+  }
+
+  @Test
+  void verifyCachedGVRIsNotInitializedWhenDBExceptionHappens(final Jdbi jdbi) {
+    // set up exception to be thrown in first call and call real method in second call.
+    doThrow(new RuntimeException("DB Not Available"))
+        .doCallRealMethod()
+        .when(metadataDao)
+        .findGenesisValidatorsRoot(any());
+
+    final GenesisValidatorRootValidator gvrValidator =
+        new GenesisValidatorRootValidator(jdbi, metadataDao);
+    final Bytes32 gvr = Bytes32.leftPad(Bytes.of(3));
+    final Bytes32 otherGvr = Bytes32.leftPad(Bytes.of(4));
+
+    // first call should throw an exception.
+    assertThatExceptionOfType(RuntimeException.class)
+        .isThrownBy(() -> gvrValidator.checkGenesisValidatorsRootAndInsertIfEmpty(gvr));
+
+    // second call should call find (which returns empty result) and insert the supplied gvr.
+    // Also cache the supplied gvr in-memory.
+    gvrValidator.checkGenesisValidatorsRootAndInsertIfEmpty(gvr);
+
+    // subsequent calls should not interact with the database
+    assertThat(gvrValidator.checkGenesisValidatorsRootAndInsertIfEmpty(gvr)).isTrue();
+    assertThat(gvrValidator.checkGenesisValidatorsRootAndInsertIfEmpty(otherGvr)).isFalse();
+
+    // verify interactions
+    verify(metadataDao, times(2)).findGenesisValidatorsRoot(any());
+    verify(metadataDao, times(1)).insertGenesisValidatorsRoot(any(), any());
   }
 
   @Test
   void checkGenesisValidatorReturnsTrueForExistingGVR(final Jdbi jdbi, final Handle handle) {
     final GenesisValidatorRootValidator gvrValidator =
-        new GenesisValidatorRootValidator(jdbi, new MetadataDao());
+        new GenesisValidatorRootValidator(jdbi, metadataDao);
     Bytes32 genesisValidatorsRoot = Bytes32.leftPad(Bytes.of(3));
     insertGvr(handle, genesisValidatorsRoot);
     assertThat(gvrValidator.checkGenesisValidatorsRootAndInsertIfEmpty(genesisValidatorsRoot))
@@ -48,7 +126,7 @@ class GenesisValidatorRootValidatorTest {
   @Test
   void checkGenesisValidatorReturnsFalseForDifferentGVR(final Jdbi jdbi, final Handle handle) {
     final GenesisValidatorRootValidator gvrValidator =
-        new GenesisValidatorRootValidator(jdbi, new MetadataDao());
+        new GenesisValidatorRootValidator(jdbi, metadataDao);
     Bytes32 genesisValidatorsRoot = Bytes32.leftPad(Bytes.of(3));
     insertGvr(handle, genesisValidatorsRoot);
     assertThat(
@@ -59,7 +137,7 @@ class GenesisValidatorRootValidatorTest {
   @Test
   void genesisValidatorRootExistsReturnsFalseWhenMetadataIsEmpty(final Jdbi jdbi) {
     final GenesisValidatorRootValidator gvrValidator =
-        new GenesisValidatorRootValidator(jdbi, new MetadataDao());
+        new GenesisValidatorRootValidator(jdbi, metadataDao);
     assertThat(gvrValidator.genesisValidatorRootExists()).isFalse();
   }
 
@@ -67,7 +145,7 @@ class GenesisValidatorRootValidatorTest {
   void genesisValidatorRootExistsReturnsTrueWhenMetadataIsNotEmpty(
       final Jdbi jdbi, final Handle handle) {
     final GenesisValidatorRootValidator gvrValidator =
-        new GenesisValidatorRootValidator(jdbi, new MetadataDao());
+        new GenesisValidatorRootValidator(jdbi, metadataDao);
     insertGvr(handle, Bytes32.leftPad(Bytes.of(3)));
     assertThat(gvrValidator.genesisValidatorRootExists()).isTrue();
   }