+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Features Added
 - Optional Azure bulk loading tags support using cli option `--azure-secrets-tags`.
 - Support Prometheus Push Gateway Metrics [#796](https://github.com/ConsenSys/web3signer/pull/796)
+- Cache Genesis Validators Root (GVR) in-memory on first database lookup. This would eliminate further database lookups 
+for GVR during sign operations and improve their performance. [#600](https://github.com/ConsenSys/web3signer/issues/600)
+
 
 ### Bugs Fixed
 - Upgrade jackson and vertx to upgrade snakeyaml to 2.0 to fix CVE-2022-1471